extract utils package from mojo+parameters
