+++ src/main/java/org/spdx/maven/utils/LicenseMapperException.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.spdx.maven;
+package org.spdx.maven.utils;
 
 /**
- * Exceptions for the collection of SPDX information
+ * Errors related to mapping Maven licenses to SPDX licenses
  *
  * @author Gary O'Neall
  */
-public class SpdxCollectionException extends Exception
+public class LicenseMapperException extends Exception
 {
 
     /**
@@ -31,15 +31,15 @@ public class SpdxCollectionException extends Exception
     /**
      *
      */
-    public SpdxCollectionException()
+    public LicenseMapperException()
     {
-        super();
+        // default constructor
     }
 
     /**
      * @param message
      */
-    public SpdxCollectionException( String message )
+    public LicenseMapperException( String message )
     {
         super( message );
     }
@@ -47,7 +47,7 @@ public class SpdxCollectionException extends Exception
     /**
      * @param cause
      */
-    public SpdxCollectionException( Throwable cause )
+    public LicenseMapperException( Throwable cause )
     {
         super( cause );
     }
@@ -56,8 +56,9 @@ public class SpdxCollectionException extends Exception
      * @param message
      * @param cause
      */
-    public SpdxCollectionException( String message, Throwable cause )
+    public LicenseMapperException( String message, Throwable cause )
     {
         super( message, cause );
     }
+
 }
