+++ src/main/java/org/spdx/maven/utils/SpdxCollectionException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright Source Auditor Inc.
+ * Copyright 2014 Source Auditor Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,44 +13,51 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.spdx.maven;
+package org.spdx.maven.utils;
 
 /**
- * Exceptions related to the LicenseManager
+ * Exceptions for the collection of SPDX information
  *
  * @author Gary O'Neall
  */
-public class LicenseManagerException extends Exception
+public class SpdxCollectionException extends Exception
 {
 
     /**
      *
      */
-    private static final long serialVersionUID = 1672757028355331818L;
+    private static final long serialVersionUID = 1L;
 
     /**
-     * @param arg0
+     *
      */
-    public LicenseManagerException( String arg0 )
+    public SpdxCollectionException()
     {
-        super( arg0 );
+        super();
     }
 
     /**
-     * @param arg0
+     * @param message
      */
-    public LicenseManagerException( Throwable arg0 )
+    public SpdxCollectionException( String message )
     {
-        super( arg0 );
+        super( message );
     }
 
     /**
-     * @param arg0
-     * @param arg1
+     * @param cause
      */
-    public LicenseManagerException( String arg0, Throwable arg1 )
+    public SpdxCollectionException( Throwable cause )
     {
-        super( arg0, arg1 );
+        super( cause );
     }
 
+    /**
+     * @param message
+     * @param cause
+     */
+    public SpdxCollectionException( String message, Throwable cause )
+    {
+        super( message, cause );
+    }
 }
