+++ src/main/java/org/spdx/maven/utils/SpdxSourceFileParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.spdx.maven;
+package org.spdx.maven.utils;
 
 import java.io.File;
 import java.io.IOException;
