+++ src/test/java/org/spdx/maven/TestSpdxMojo.java
@@ -31,6 +31,7 @@ import org.spdx.library.model.license.AnyLicenseInfo;
 import org.spdx.library.model.license.ExtractedLicenseInfo;
 import org.spdx.library.model.license.LicenseInfoFactory;
 import org.spdx.library.referencetype.ListedReferenceTypes;
+import org.spdx.maven.utils.TestSpdxFileCollector;
 import org.spdx.spdxRdfStore.RdfStore;
 import org.spdx.storage.ISerializableModelStore;
 import org.spdx.storage.simple.InMemSpdxStore;
