+++ src/test/java/org/spdx/maven/utils/TestUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.spdx.maven;
+package org.spdx.maven.utils;
 
 import java.util.Arrays;
 import java.util.Collection;
